requirements.txt
anyio
asyncpg
click
colorama
databases
fastapi
greenlet
gunicorn
h11
idna
pydantic
python-dotenv
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
psycopg2-binary
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
